Add a constructor on Article entity
This also makes the properties not nullable, as it does not make sense
to have an article without a title or a body.

The id is handled by Doctrine (which is the persistence layer), so it
doesn't make sense to be able to access it uninitialized.

An uuid would have been better, but mysql being mysql...